Fix skeleton
