--- app/components/Views/confirmations/components/custom-amount-info-skeleton/custom-amount-info-skeleton.tsx
@@ -1,19 +1,0 @@
-import React from 'react';
-import { PayTokenAmountSkeleton } from '../pay-token-amount';
-import { PayWithRowSkeleton } from '../rows/pay-with-row';
-import InfoSection from '../UI/info-row/info-section';
-import { CustomAmountSkeleton } from '../transactions/custom-amount';
-import { DepositKeyboardSkeleton } from '../deposit-keyboard';
-
-export function CustomAmountInfoSkeleton() {
-  return (
-    <>
-      <CustomAmountSkeleton />
-      <PayTokenAmountSkeleton />
-      <InfoSection>
-        <PayWithRowSkeleton />
-      </InfoSection>
-      <DepositKeyboardSkeleton />
-    </>
-  );
-}
