--- app/components/Views/confirmations/components/custom-amount-info-skeleton/index.ts
@@ -1,1 +1,0 @@
-export * from './custom-amount-info-skeleton';
